Add initial objects
This starts the basic layout of the project. I'm interested in watching
the basic stats we see for given protocols and the repository. To
fufull this I've added a Bandwith Chart "BWChart" to track total
bandwith usage over time. I've also added a Bandwith Table "BWTable"
to track bandwith usage by protocol. Finally, there is a Repository
Paragrap "RepoPar" that describes the current state of the repository.